shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ThuyVan032004", email = "dovan032004@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.6",
    "feast>=0.47.0",
    "nltk>=3.9.2",
    "pandas<3.0.0",
    "pendulum>=3.1.0",
    "pydantic>=2.12.5",
    "seaborn>=0.13.2",
    "wordcloud>=1.9.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
